// src/app/api-client/model/apiResponseObject.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ApiResponseObject { 
    success?: boolean;
    data?: object;
    message?: string;
}